+++ doc/CHANGELOG.md
@@ -22,6 +22,7 @@
 - Support for World of Darkness Homebrew 10s cancel 1s
 - Support for Vampire 5e
 - Support for Laser and Feelings
+- Support Level up D&D 5th Edition
 - "stability and performance improvements"
 
 ## [1.3.0] - 2025-07-03